added branching, permissions

+++ mortar/views.py
@@ -1,11 +1,14 @@
-import time
+import time, os
 from django.forms import forms
 from django.views.generic import TemplateView, ListView, FormView
 from django.utils.text import slugify 
 from django.urls import reverse
-from django.http import HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect
 from django.db import transaction
+from django.db.models import Q
 from django.shortcuts import render
+from django.contrib.auth import get_user_model
+from django.contrib.auth.mixins import LoginRequiredMixin
 from xml.etree import ElementTree as etree
 from django_mptt_admin.admin import DjangoMpttAdminMixin
 from django_mptt_admin.util import *
@@ -13,15 +16,20 @@ from mptt.templatetags.mptt_tags import cache_tree_children
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import authentication, permissions
+from wsgiref.util import FileWrapper
+import mimetypes
 from .models import Project, ProjectTree, Category
 from .forms import TreeForm, ImportForm, CategoryForm
 import re
 
-class ProjectListView(ListView):
+class ProjectListView(ListView, LoginRequiredMixin):
     model = Project
     context_object_list = 'project_list'
 
-class ProjectDetailView(FormView):
+    def get_queryset(self):
+        return Project.objects.filter(assigned=self.request.user)
+
+class ProjectDetailView(FormView, LoginRequiredMixin):
     template_name = 'mortar/project_detail.html'
     form_class = TreeForm
 
@@ -29,25 +37,28 @@ class ProjectDetailView(FormView):
         context = self.get_context_data(**kwargs)    
         trees = context['project_trees']
         for tree in trees:
-            if request.GET.get(str(tree.pk) + ('-duplicate')) == "True":
+            if request.GET.get(str(tree.pk) + ('-duplicate')) == "True" and request.user in context['project_users']:
                 new_tree = ProjectTree(
                     name=tree.name + " (copy)",
                     slug=slugify(tree.name+" (copy)"),
+                    owner=request.user,
                     project=context['project']
                 )
                 new_tree.save()
                 self.copy_nodes(tree, new_tree)
 
-            elif request.GET.get(str(tree.pk) + ('-delete')) == "True":
+            elif request.GET.get(str(tree.pk) + ('-delete')) == "True" and request.user == tree.owner:
                 tree.delete()
 
-        context['project_trees'] = context['project'].trees.all()
+        context['project_trees'] = ProjectTree.objects.filter(project=context['project'])
         return self.render_to_response(context)
     
     def get_context_data(self, *args, **kwargs):
         context = super(ProjectDetailView, self).get_context_data(**kwargs)
+        context['user'] = self.request.user
         context['project'] = Project.objects.get(slug=self.kwargs.get('slug'))
-        context['project_trees'] = context['project'].trees.all()
+        context['project_users'] = context['project'].assigned.all()
+        context['project_trees'] = ProjectTree.objects.filter(project=context['project'])
         return context
 
     def form_valid(self, form, *args, **kwargs):
@@ -55,6 +66,7 @@ class ProjectDetailView(FormView):
         new_tree = ProjectTree(
             name=form.cleaned_data['name'],
             slug=slugify(form.cleaned_data['name']),
+            owner=self.request.user,
             project=context['project']
         )
         new_tree.save()
@@ -83,24 +95,26 @@ class ProjectDetailView(FormView):
                 )    
                 new_node.save()
 
-class TreeDetailView(TemplateView):
+class TreeDetailView(TemplateView, LoginRequiredMixin):
     template_name = 'mortar/tree_detail.html'
     selected_node = None
 
     def get_context_data(self, *args, **kwargs):
         context = super(TreeDetailView, self).get_context_data(**kwargs)
         tree = ProjectTree.objects.get(slug=self.kwargs.get('slug'))
+        context['user'] = self.request.user
         context['tree'] = tree
         context['tree_json_url'] = reverse('tree-json', kwargs={'slug': tree.slug})
         context['insert_at_url'] = reverse('tree-insert', kwargs={'project_slug': self.kwargs.get('project_slug'), 'slug': tree.slug})
+        context['branch_url'] = reverse('tree-branch', kwargs={'project_slug': self.kwargs.get('project_slug'), 'slug': tree.slug})
         context['edit_url'] = reverse('tree-edit', kwargs={'project_slug': self.kwargs.get('project_slug'), 'slug': tree.slug})
         context['app_label'] = "mortar"
         context['model_name'] = "category"
         context['tree_auto_open'] = 'true'
         context['autoescape'] = 'true'
         context['use_context_menu'] = 'false'
-        #context['paths'] = self.get_all_paths(tree)
         context['importform'] = ImportForm(self.request.POST, self.request.FILES)
+        print(context)
         return context
 
     def post(self, request, *args, **kwargs):
@@ -112,13 +126,31 @@ class TreeDetailView(TemplateView):
             position = request.POST.get('position')
             target_instance = Category.objects.get(pk=target_id)
             self.move_node(instance, position, target_instance)
+        elif request.POST.get('export'):
+            return self.tree_to_csv()
         elif form.is_valid():
             if request.FILES['file'].name.split('.')[-1] == 'mm':
                 self.read_mindmap(context['tree'], request.FILES['file'].read())
             else:
                 self.read_csv(context['tree'], request.FILES['file'].read())
-
         return render(request, self.template_name, context=self.get_context_data(**kwargs))
+
+    def tree_to_csv(self):
+        context = self.get_context_data()
+        tree = context['tree']
+        filename = tree.slug + ".csv"
+        with open(filename, "w") as f:
+            f.write("fullPathName,name,regex\n")
+            for cat in Category.objects.filter(projecttree=tree, is_rule=True):
+                f.write(",".join([cat.full_path_name, cat.name, cat.regex]) + "\n")
+        f = open(filename, 'rb')
+        wrapper = FileWrapper(f)
+        mt = mimetypes.guess_type(filename)[0]
+        response = HttpResponse(wrapper, content_type=mt)
+        response['Content-Length'] = os.path.getsize(filename)
+        response['Content-Disposition'] = 'attachment; filename=%s' % filename
+        response['X-Sendfile'] = os.path.realpath(filename)
+        return response    
 
     @transaction.atomic()
     def move_node(self, instance, position, target_instance):
@@ -141,15 +173,19 @@ class TreeDetailView(TemplateView):
         paths = self.get_all_paths(tree)
         new_rules = self.preprocess_csv(csvfile)
         created = 0
-        with transaction.atomic():
-            for key in new_rules.keys():
-                print("Processing " + str(len(upload)) + " new records")
-                parent, all_paths = self.create_categories(key, tree, paths)
-                for rule in new_rules[key]:
-                    self.create_rule(parent, tree, rule['name'], rule['regex'])
-                    created += 1
-                    print(str(created) + "/" + str(len(upload)) + " records processed")
-                paths = all_paths
+        for key in new_rules.keys():
+            print("Processing " + str(len(upload)) + " new records")
+            parent, all_paths = self.create_categories(key, tree, paths)
+            #to_create = []
+            for rule in new_rules[key]:
+                #to_create.append(Category(parent=parent, projecttree=tree, name=rule['name'], is_rule=True, regex=rule['regex']))
+                self.create_rule(parent, tree, rule['name'], rule['regex'])
+                created += 1
+
+            #with transaction.atomic():
+            #    Category.objects.bulk_create(to_create)
+                print(str(created) + "/" + str(len(upload)) + " records processed")
+            paths = all_paths
 
     def preprocess_csv(self, csvfile):
         processed = {}
@@ -163,25 +199,8 @@ class TreeDetailView(TemplateView):
                     processed[csv[0]].append({"name": csv[1], "regex": csv[2]})
         return processed
 
-    def unroll_csv(self, csvfile):
-        unrolled = {}
-        upload = csvfile.decode().split('\n')[1:]
-        for line in upload:
-            csv = line.strip('\r').split(',')
-            if len(csv) == 3:
-                path = csv[0].split('.')
-                for x in range(0, len(path)):
-                    pass
-#    def rule_exists(self, tree, name, regex, path):
-#        rules = Category.objects.filter(projecttree=tree, is_rule=True, name=name, regex=regex)
-#        for rule in rules:
-#            print("%s %s" % (rule.full_path_name, path))
-#            if rule.full_path_name == path:
-#                return True
-#        return False
-    
     def get_all_paths(self, tree):
-        leaves = Category.objects.filter(children__isnull=True, projecttree=tree)
+        leaves = Category.objects.filter(projecttree=tree)
         paths = []
         for leaf in leaves:
             path = leaf.full_path_name
@@ -196,20 +215,12 @@ class TreeDetailView(TemplateView):
             filtered = [x for x in tree.categories.all() if x.full_path_name == new_path]
             if len(filtered) > 0:
                 return filtered[0],all_paths
-        # TODO see if any part of new_path exists
-        #for path in all_paths:
-        #    if new_path.startswith(path):       
-        #        a = new_path.split('.')
-        #        b = path.split('.')
-        #        c = set(a) - set(b)
-        #        cats = [cat for cat in a if cat in c]
-        #        print("Set Difference")
-        #        print(cats)
 
         # create any categories that need creating
         last_cat = None
         for cat in cats:
-            new_cat,created = Category.objects.get_or_create(projecttree=tree, name=cat, parent=last_cat)
+            with transaction.atomic():
+                new_cat,created = Category.objects.get_or_create(projecttree=tree, name=cat, parent=last_cat)
             print(new_cat)
             if created:
                 print("New Category: %s" % cat)
@@ -233,7 +244,7 @@ class TreeDetailView(TemplateView):
             print("%s: Invalid Regex" % name)
             raise forms.ValidationError("Invalid Regex")
 
-class TreeJsonApi(APIView):
+class TreeJsonApi(APIView, LoginRequiredMixin):
         
     def get(self, request, format=None, **kwargs):
         tree = ProjectTree.objects.get(slug=self.kwargs.get('slug'))
@@ -275,12 +286,13 @@ class TreeJsonApi(APIView):
         return tree
                     
 
-class TreeInsertView(FormView):
+class TreeInsertView(FormView, LoginRequiredMixin):
     form_class = CategoryForm
     template_name = 'mortar/tree_insert.html'
 
     def get_context_data(self, *args, **kwargs):
         context = super(TreeInsertView, self).get_context_data(**kwargs)
+        context['user'] = self.request.user
         context['project_slug'] = self.kwargs.get('project_slug')
         context['tree'] = ProjectTree.objects.get(slug=self.kwargs.get('slug'))
         context['insert_at'] = Category.objects.get(id=self.kwargs.get('id'))
@@ -302,12 +314,43 @@ class TreeInsertView(FormView):
         node.save()
         return HttpResponseRedirect(reverse('tree-detail', kwargs={'project_slug':context['project_slug'], 'slug':context['tree'].slug}))
 
-class TreeEditView(FormView):
+class TreeBranchView(APIView, LoginRequiredMixin):
+    def get(self, request, *args, **kwargs):
+        project = Project.objects.get(slug=self.kwargs.get('project_slug'))
+        tree = ProjectTree.objects.get(slug=self.kwargs.get('slug'))
+        root = Category.objects.get(id=self.kwargs.get('id'))
+        branch = root.get_descendants(include_self=True)
+        new_tree = ProjectTree.objects.create(
+            name='Branch of ' + str(root.id),
+            slug=slugify('Branch of ' + str(root.id)),
+            owner=request.user,
+            project=project,
+        )
+
+        for node in branch:
+            parent = None
+            if node.parent and node != root:
+                possibles = Category.objects.filter(projecttree=new_tree, name=node.parent.name)
+                parent = [p for p in possibles if p.full_path_name in node.parent.full_path_name]
+                parent = parent[0]
+            new_node = Category.objects.create(
+                name=node.name,
+                parent=parent,
+                projecttree=new_tree,
+                is_rule=node.is_rule,
+                regex=node.regex
+            )                                                                   
+            new_node.save()
+ 
+        return HttpResponseRedirect(reverse('tree-detail', kwargs={'project_slug': project.slug, 'slug': new_tree.slug}))
+
+class TreeEditView(FormView, LoginRequiredMixin):
     form_class = CategoryForm
     template_name = 'mortar/tree_insert.html'
 
     def get_context_data(self, *args, **kwargs):
         context = super(TreeEditView, self).get_context_data(**kwargs)
+        context['user'] = self.request.user
         context['project_slug'] = self.kwargs.get('project_slug')
         context['tree'] = ProjectTree.objects.get(slug=self.kwargs.get('slug'))
         context['edit'] = Category.objects.get(id=self.kwargs.get('id'))
@@ -331,3 +374,4 @@ tree_detail = TreeDetailView.as_view()
 tree_json = TreeJsonApi.as_view()
 tree_insert = TreeInsertView.as_view()
 tree_edit = TreeInsertView.as_view()
+tree_branch = TreeBranchView.as_view()
